+++ packages/wethegit-components-cli/src/commands/add/utils/promptForComponents.ts
@@ -1,17 +1,23 @@
 import prompts from "prompts";
 import chalk from "chalk";
 
-import COMPONENTS_INDEX from "../../../component-index";
+import { REGISTRY_INDEX } from "../../../registry-index";
 
-export function promptForComponents() {
+export function promptForComponents(): Promise<{
+  selectedComponentNames: (keyof typeof REGISTRY_INDEX)[];
+  proceed: boolean;
+}> {
   return prompts(
     [
       {
         type: "multiselect",
         name: "selectedComponentNames",
         message: "What components would you like to add?",
-        choices: Object.values(COMPONENTS_INDEX)
-          .filter(({ dontShowOnPrompt }) => !dontShowOnPrompt)
+        choices: Object.values(REGISTRY_INDEX)
+          .filter(
+            ({ dontShowOnPrompt, type }) =>
+              type === "component" && !dontShowOnPrompt
+          )
           .map(({ name }) => ({
             title: name,
             value: name,