+++ packages/wethegit-components/src/utilities/fixed-forward-ref/fixed-forward-ref.ts
@@ -1,0 +1,7 @@
+import { forwardRef } from "react";
+
+type FixedForwardRef = <T, P = object>(
+  render: (props: P, ref: React.Ref<T>) => React.ReactNode
+) => (props: P & React.RefAttributes<T>) => JSX.Element;
+
+export const fixedForwardRef = forwardRef as FixedForwardRef;