+++ packages/wethegit-components-cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @wethegit/components-cli
 
+## 2.4.2
+
+### Patch Changes
+
+- Fix recursive typescript transformation. Fix prettier check. Ties config and regitry with types. Adds Tag component for generic components.
+- Updated dependencies
+  - @wethegit/components@2.2.2
+
 ## 2.4.1
 
 ### Patch Changes